+++ src/main/java/com/samaxes/maven/minify/plugin/MinifyMojo.java
@@ -23,7 +23,10 @@ package com.samaxes.maven.minify.plugin;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.concurrent.*;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
@@ -159,7 +162,7 @@ public class MinifyMojo extends AbstractMojo {
      * encoding types.
      * </p>
      *
-     * @parameter expression="${charset}"
+     * @parameter expression="${charset}" default-value="UTF-8"
      * @since 1.3.2
      */
     private String charset;
@@ -215,14 +218,6 @@ public class MinifyMojo extends AbstractMojo {
     private int bufferSize;
 
     /**
-     * Maximum execution time in seconds.
-     *
-     * @parameter expression="${timeout}" default-value="30"
-     * @since 1.5
-     */
-    private long timeout;
-
-    /**
      * Show source file paths in log output.
      *
      * @parameter expression="${debug}" default-value="false"
@@ -267,6 +262,7 @@ public class MinifyMojo extends AbstractMojo {
     /**
      * Executed when the goal is invoked, it will first invoke a parallel lifecycle, ending at the given phase.
      */
+    @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
         if (skipMerge && skipMinify) {
             getLog().warn("Both merge and minify steps are configured to be skipped.");
@@ -285,19 +281,16 @@ public class MinifyMojo extends AbstractMojo {
         ExecutorService executor = Executors.newFixedThreadPool(2);
         try {
             List<Future<Object>> futures = executor.invokeAll(processFilesTasks);
-            executor.shutdown();
-            executor.awaitTermination(timeout, TimeUnit.SECONDS);
-            for (Future<Object> future : futures)
-            {
-                try
-                {
+            for (Future<Object> future : futures) {
+                try {
                     future.get();
-                } catch (ExecutionException e)
-                {
+                } catch (ExecutionException e) {
                     throw new MojoFailureException(e.getMessage(), e);
                 }
             }
+            executor.shutdown();
         } catch (InterruptedException e) {
+            executor.shutdownNow();
             throw new MojoFailureException(e.getMessage(), e);
         }
     }